// src/app/shared/models/receipt.model.ts
export class Receipt {
  id: string = '';
  fileName: string = '';
  type: string = '';
  folderName: string = '';
  companyName: string = '';
  siren: string = '';
  address: string = '';
  date: string = '';
  client: string = '';
  exclTax: string = '';
  vatRate: string = '';
  vat: string = '';
  inclTax: string = '';
  imgSrc: string = '';
}